pyproject.toml
[project]
name = "it-support-agent"
version = "0.1.0"
description = "RAG-powered IT support agent using Strands SDK and Amazon Bedrock"
requires-python = ">=3.11"
dependencies = [
    "strands-agents>=0.1.0",
    "strands-agents-tools>=0.1.0",
    "faiss-cpu>=1.7.4",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "boto3>=1.35.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "pytest-cov>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/it_support_agent"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
